+++ src/App.vue
@@ -2,7 +2,15 @@
   <div id="app">
     <h3 class="mt-5">eventzimmer</h3>
     <button type="button" @click="backToTop" id="backToTop" class="btn btn-secondary btn-back-to-top fixed-bottom mb-2" style="display: none"><i class="fas fa-angle-up"></i></button>
-    <cities></cities>
+    <div class="row mb-1 mt-1">
+      <div class="col-sm"></div>
+      <div class="col-sm-2">
+        <categories></categories>
+      </div>
+      <div class="col-sm-2">
+        <cities></cities>
+      </div>
+    </div>
     <events></events>
   </div>
 </template>
@@ -10,12 +18,14 @@
 <script>
 import Events from '@/components/Events'
 import Cities from '@/components/Cities'
+import Categories from '@/components/Categories'
 
 export default {
   name: 'app',
   components: {
     Events,
-    Cities
+    Cities,
+    Categories
   },
   mounted () {
     window.onscroll = () => {